Remove code "converting" T_PERM{2,4} to Julia arrays
This code was untested, and seems of dubious use to me: Instead, if somebody
needs to do this, they should be able to simply use `ListPerm`, followed by a
conversion from plist to array.

If that is not fast enough because, say, someone need to convert a million
permutations, then with overwhelming probably they are doing something very
wrong. But if a legitimate use case pops up, we can still add back similar
code. Although I still would not want to have it in the "generic"
`ConvertedToJulia` function; rather it should be in a dedicated
`ConvertGAPPermToJuliaArray` function.