Replace GAP orbit implementation
The new one is much faster. Indeed, for me it is faster than the Julia
version. However, that also could be optimized; the main reason I did not yet
do that is that one needs to access a hash function for the orbit elements
(here: GAP integers), and I did not yet bother to figure out how to do that.