Remove GAPRat code
This code is not needed anymore due the existence of the convenience wrappers
for arithmetic operations on GAP objects in Julia, resp. on Julia objects in
GAP.

Moreover, this code was buggy; and was not yet converted to use
`ForeignGAP.MPtr` (leading to GC crash). All in all, it seems best to
"conclude" this experiment by removing it.

Resolves #71