WIP: Import conversion rules into AutoDoc
Problem: AutoDoc only supports a subset of Markdown, and in particular does
not support those (non-standard) Markdown tables.

So it might be better to either convert this manually into GAPDoc XML; or to
import this into the "Julia documentation", however that is done, as that
seems to support proper Markdown (but I don't know if it supports tables).