show GAP help in Julia with 'GAP.show_GAP_help'
This is a first approximation.

The functions in the file `helpstring.g` arose from those
in GAP's library file `helpbase.gi`.
The idea is to extract the requested information
from the text format files of the GAP manuals,
and to collect it in a string,
also in the case where several matches are found.

Currently the following problems are known.

- When the last subsection of a section is requested
  then also the part of the next section (in the same chapter)
  until the start of its first subsection is shown;
  apparently the sections themselves have no entry numbers
  in the sense of `HELP_BOOK_HANDLER.(book.handler).HelpData`.

  (An example is `SizeScreen`.)

- Some pieces of GAP help are very long,
  an example is the entry for `Size` in the HAP package.

On the other hand, duplicate entries are filtered out,
which should perhaps also done in GAP's help system.