added some tests for help requests
The tests cover the situation where one asks for
help about a topic, and where an exact match is found
or just partial matches are found or no matches are found at all.

There is code in `helpstring.g` which cannot be reached via these requests,
such as requests for the next or previous topic in a manual
(which do not make sense in the current setup for Julia),
but I wanted to have a complete "string returning" variant
of GAP's `helpbase.gi`.