libjulia ignores .julia/config/startup.jl ...
..., let GAP load it if wanted.

The changes adjust pull request #19 to the behaviour of the current Julia.
Note that Julia uses a different startup file already for quite some time.

Changing the name of the GAP user preference might be regarded as not nice,
but I think it is o.k. since there was no documentation.

This pull request addresses part of issue #284.