catch GAP error messages in a Julia session
Up to now, GAP (running in a Julia session) printed its error messages
to the screen.
Now GAP's `ERROR_OUTPUT` gets redirected to a separate stream,
which is read and reset by the `error_handler` function in Julia.

(I interpret the failed tests in the first attempt in such a way
that this code does not work in Julia 1.1, but that it can be used
in later Julia versions.)