Remove special support for using a self-compiled GAP
This is not something we want to encourage in general, and it's only
really useful for GAP.jl developers; and even that only rarely: when
testing against GAP master, for example, which we might want to do again
in the future. But then we can use artifact overrides, or simply copy /
symlink the required files into the right place.