HACK: use custom GAP 4.11.0 tarball
The official GAP 4.11.0 tarballs contains `._*` files created by Apple's
version of tar. Unfortunately this means that the directory tree
resulting from this archive differs depending on whether one used a
regular BSD or GNU tar; or used one of Apples' version of tar.
Unfortunately this means the git-tree-sha1 differs.

To avoid this, we uploaded our own modified version of the 4.11.0
tarball with the only change that those `._*` files are all removed.