Travis: import Julia pkgs in install step, abort on error
Change the install step so that we notice such breakage earlier, by
immediately `using` the packages we just installed. This way, if there
are any trouble with importing the packages, we'll find out about these
much earlier. As a bonus, the packages are already precompiled when we
run the tests later on.

Also abort the Travis job immediately when such an error occurs.