proposed fix for a conversion inconsistency
Currently the following happens.
```
gap> x:= JuliaEvalString( "2^60" );
1152921504606846976
gap> GAPToJulia( 2^60 );
<Julia: 1152921504606846976>
```
The number `x` is not an immediate integer, and `Julia.Base.sqrt( x )` yields an error.
The result of `GAPToJulia` (computed using `gap_to_julia`) looks correct.
`JuliaEvalString` calls `gap_julia`, which calls `ObjInt_Int8` in this case;
I propose to use `NewJuliaObj` instead.

After this change, the tests exhibit a problem with the method
```
julia_to_gap(x::Int64)  = x
```
which is simply wrong.

Perhaps this proposal is not a good idea.
Converting `Int64` to GAP becomes more expensive because the argument has to be checked.
For the moment, I have no good idea to improve this.