Start GAP via Julia
This was borne out of need but now actually seems quite nice...

The need was that in order to use JLLs, we need to link GAP against libraries
in various prefixes; prefixes which depend on artifacts, and thus on the hash
codes of those artifacts, which can and do change as the JLLs get updated.

So in order to load the "regular" GAP binary, we'd have to set OS specific
environment variables to point the dynamic linker at the right lib
directories. But the values for those variables can change, so we can't just
put them into the generated gap.sh file! We need to query Julia for them...
but to do that, we need to load those JLLs. But we can't just do that via
`using`, as they might not have been added to the current environment! The
only thing we can reliably "use" is GAP.jl itself. But if we do that, we are
effectively loading a full copy of GAP already! So if we go that way, we
immediately *double* the startup time of GAP.

So the logical next step is: why not "use" the GAP we are loading anyway?
This way, we also don't need to mess with passing flags to the dynamic linker
via environment variables; Julia takes care of all of that for us!