Workaound issue in GAP profiling package
The mtime of the `configure` script of the GAP package "profiling" inside the
GAP 4.11.0 tarball unfortunately is older than the mtime of one of its source
files; moreover, maintainer mode is on. As a result, trying to build that
package results in autoconf being invoked to regenerate the `configure`
script. But that fails if autoconf is not installed.

But there really is no need to regenerate `configure` in the first place; so
we just `touch` it to stop it from being regenerated carelessly.