Rename artifact gap-4-11-0 to gap
When I introduce this, I somehow thought that it made sense or even was
necessary to include the GAP version in the artifact name, but really,
it isn't. Dropping it slightly simplifies our code, but more
importantly, it will make it a tad easier to maintain CI scripts that
override the artifact (as they won't have to be adjusted to match
changes in the GAP version we bundle).