added documentation and doctests for constructors ...
... and had to adjust the code, as follows:

- Fixed the `Char(obj::GapObj)` method,
- removed the `Vector{UInt8}(obj::GapObj)` method, it was covered already by the `Vector{T}(obj::GapObj)` method,
- admit `gap_to_julia(Vector{T}, obj)` also if `obj` is in `Globals.IsVectorObj`, added tests,
- fixed `gap_to_julia(StepRange, obj)`; note that `StepRange(obj)` is not supported (but `StepRange{Int,Int}(obj)` is), now we call `convert(T, obj)` instead of `T(obj)` (thanks to @rfourquet for a discussion about this),
- support `gap_to_julia(obj)` for `obj` in `Globals.IsMatrixObj`,
- added tests for `Matrix{T}(obj)` where `obj` is in Globals.IsMatrixObj`.

Perhaps we want to change the behaviour of `gap_to_julia(obj)` for `obj` in `Globals.IsMatrixObj`:
Up to now, there was no such method (without prescribed Julia type).
Currently `Globals.IsList` is checked before `Globals.IsMatrixObj`,
which means that a conversion of a GAP matrix to a nested Julia array
is preferred.
If we would check for `Globals.IsMatrixObj` first then the default
conversion of GAP matrices would produce 2-dimensional Julia arrays.