support computed external references
a clean solution to compute external references

Preliminary support for computed external references in the documentation
of GAP.jl had been proposed in pull request #472,
but this was of course not satisfactory because it misused Documenter.jl.

In the discussion of JuliaDocs/Documenter.jl/issues/1343,
@mortenpi suggested to get the required functionality by adding a plugin
to Documented.jl, analogous to an implementation of citations by @ali-ramadhan.
This works well, and from my point of view the issue JuliaDocs/Documenter.jl/issues/1343 can be closed.

Thanks a lot to @mortenpi for the hint, and to @ali-ramadhan for the code.

The details of the changes are:

- added a plugin for Documenter.jl
- replaced the GAP function `UrlOfManualEntry` by `MatchURLs`
  (thanks to @frankluebeck for this code)
- removed the now unnecessary GAP function `ComputeLinksToGAPManuals` and the Julia function `compute_links_to_gap_manuals`