changed the `show` method for GAP objects ...
... such that not more lines are shown than fit on the screen
(including the Julia prompts), by omitting the middle part
of too long `ViewObj` results.

The unicode symbol `⋮` is used, which seems to be standard in Julia
for this purpose.

I hope this does not conflict with the "no unicode" policy
that is mentioned in the discussion about Nemocas/AbstractAlgebra.jl/pull/389.
(Where is this policy stated?)