two-step help request via a menu
This is a draft proposal of an alternative access to documentation.
Instead of showing all matching entries for a help request
(which is Julia's way to look at things), a menu with an overview
of the matching entries is shown, from which one can choose what
one wants to see.
I think this is more useful than looking through several pages of
screen output.

A problem is how to test this kind of help access, since one has to enter
some characters (at least the `q` for exiting the menu) in interactive mode.
I have tried to simulate a terminal, but this is too simpleminded,
since one gets warnings when the tests are run.

(If this kind of help access is interesting then it would perhaps
make sense to provide also a variant that is based on GAP's Browse package
instead of Julia's REPL menus.)