Partial fix for dlsym() not finding symbols.
This change deals with dlsym() lookups failing in GAP.jl. Rather
than searching the global namespace, we're now looking them up
specifically in the library.

This is still broken on macOS, as macOS loads two different
instances of JuliaInterface.so. The current workaround (see
comments in src/cbind.jl) works only for Julia >= 1.4.