try to make the changed version work with Julia 1.3
The changed assignment to `Globals.AlwaysPrintTracebackOnError` moves the error from the initialization to the situation where a GAP error occurs at runtime.

The `Base.invokelatest` call in the error handler lets the runtime error disappear.

(Taking into account which trivial change caused this problem with Julia 1.3, it is surprising that we did not meet more instances of it.)