moved the `g_str` macro code to a helper function
in order to admit testing the error handling.

My original intention was to have a test that notices when the behaviour of the macro changes.
Note that the input `g"abc\\def"` leads to the string `"abc\\\\def"` (length 8) inside the macro, whereas the input `g"abc\\"` leads to the string `"abc\\"` (length 4) inside the macro.
(Isn't this a Julia bug?)

The helper function takes already the wrong input as its argument.
Testing that this runs into an error is meaningful only in the sense that the code line in question is covered.