optimize Vector{T} constuctor by avoiding elmlist variable
The `elmlist` variable was storing either the `ElmList` or the
`Globals.ELM_LIST` object, and was called later as `elmlist(obj, i)`.
But unfortunately, this is not (yet) a pattern that the Julia compiler
optimizes well. So instead, just write explicitly the calls to `ElmList`
and `ELM_LIST`, with a branch.