fix `GAP.Packages.load`
I think this code does not what one wants (my fault).

If the package cannot be loaded and the keyword argument `install` is `true`
then it does not help to check whether an admissible version is already
known to GAP, since the package directory may be there but the package
has not been compiled yet.
(Besides that, the check for the availability of the package was not correct.)
Thus `GAP.Packages.install` should always be called in this case.

After this fix is available,
the call `GAP.Packages.install("ferret")` in `Oscar.jl/examples/GaloisGrp.jl`
should be replaced by `GAP.Packages.load("ferret", install = true)`,
otherwise a new installation is forced at runtime whenever the file gets read.