Improve _GAP_TO_JULIA
Reimplement more of the julia_gap logic; this makes the cases we now handle on
the Julia side potentially a bit faster (??? untested) but most importantly it
means this code now works even if JuliaInterface is not yet loaded. On the
downside, if we end up ccalling julia_gap anyway, it performs a bunch of
checks a second time.