Remove JuliaBindCFunction
This was a very strange function, and was difficult to use correctly
without a crash, and also very limited in its applications.

It was mainly meant to provide a performance boost; but I think for
real world applications, we need a different approach which takes
Julia's multiple dispatch and argument type specialization properly
into account.

Fixes #349