try to make the Julia help available in a GAP session
For that, notify a virtual GAP help book called `Julia`,
extract the Julia help about the Julia object in question
(including the relevant ANSI escape sequences),
and feed it into the data structures expected by the GAP help system.

For example, the following requests will include entries from the
Julia help.
```
?size
?Julia:size
?Julia:GAP.julia_to_gap
?Julia:GAP.julia_help_string
```

One problem is that the Julia help system is case sensitive,
but the GAP help system turns the given string to lowercase
in the beginning.
Note that Julia does not understand the request `?base.sqrt`,
one has to enter `?Base.sqrt`.