Fix potential GC related crashes
When we ccall a function to which we pass pointers (in)to Julia objects, if
that function may end up triggering a garbage collection, in general we must
use GC.@preserve on the Julia objects we passed in to ensure they are not
garbage collected. This is not needed if the Julia object is in fact a GAP
object (as our stack scanning logic takes care of those), or if other Julia
code later on is guaranteed to access the object in danger of being GCed.

This could result in a crash in practice when using `evalstr`; the string
containing GAP code could be garbage collected while it was still being
executed.

In the same vein, call `JL_GC_PUSH1(&args)` in the C function `call_gap_func`.