catch GAP errors in `GAP.evalstr`
The right way to do this is to check the return value of `evalstr_ex`,
and to call `GAP.error_handler` if necessary,
as was proposed in the discussion of gap-system/gap/issues/4199.

This is intended to resolve issue #580.

(The change does of course not affect the problems described in
issue #596.)