Don't unwrap IS_JULIA_FUNC objects automatically
... when moving them to the Julia side.

This way, Julia code can store them and pass them later on to GAP code again,
e.g. as action function for G-set related computations.

This also fits better into our philosophy regarding conversions: the principle
of only doing necessary or "obvious" AND cheap conversions. But this automatic
unwrapping was neither necessary (as long as we provide an explicit unwrap
function on the Julia side), nor necessarily cheap (as discarding the wrapper
meant that one may have to recreate many such wrappers). Moreover, it impeded
the usefulness of these wrappers.