changed the definition of `UnwrapJuliaFunc`
Up to now, we defined `UnwrapJuliaFunc(x::Function) = x` and then
added a method for `GapObj`.
This is not good enough, since there are callable Julia objects
which do not have the type `Function`.

Concretely:
`Polymake.polytope.Polytope` is a callable Julia object
which is not a `Function` and which admits keyword arguments.
When we call it from GAP via `CallJuliaFunctionWithKeywordArguments`,
we get an error with the old version.