Remove LoadPackageAndExposeGlobals
This function always was very peculiar and for special interests only.
It is also very easily abused resp. used in a problematic way.

I believe that any code which uses this right now should be rewritten
(and I'll be happy to help with that by suggesting ways to do it).
Alternatively, one can of course also copy the code deleted by this
patch into one's own code.