Remove `ImportJuliaModuleIntoGAP`
This function and its documentation were highly misleading, and as a
result a lot of code used it inappropriately. To wit, it was *not*
actually necessary to call it in order to use a Julia module in GAP;
*only* if that module was not imported into the Julia `Main` module, and
then of course only if the GAP code relied on that.

But in practice this lead to a lot of people (including myself in the
past) of writing brittle code that only worked if the user did
`Pkg.add("NAME")` for the module to be imported into GAP. But in
practical

For those who genuinely needed `ImportJuliaModuleIntoGAP("NAME")`, they
can replace this by `JuliaEvalString("import NAME")`. The only thing
missing then is that tab completion for `Julia.NAME` will not work
out of the box. We'll restore this by some other means eventually.