Better tab completion in GAP for members of Julia "record"
This restores tab completion after `Julia.`, by overwriting ("monkey
patching") the relevant GAP function. This is only safe because we use a fixed
GAP version; of course if we update GAP, we must update this patch. But once
https://github.com/gap-system/gap/pull/4667 is merged and a newer snapshot of
GAP is used by GAP.jl, we can get rid of this hack again.