Rename lib/ -> gap/
This directory contains some GAP source code. Having it in lib/ sometimes
caused problems, because the root directory of GAP.jl is also a GAP root
directory; and so any file in lib/ would shadow any file with the same name in
the GAP library. That could lead to "hilarious" confusion.