fixed `@gapattribute`
In the construction of setter and tester, we have to distinguish the Julia object (called `juliaarg`) and the corresponding GAP object (called `gaparg`).
The two names are equal in the test example, but in typical Oscar situations,
the GAP object corresponding to the Julia object `G` is `G.X`.

Perhaps the problem has not been observed before because setters and testers are not really used in Oscar.