install a `GAP.Globals.Download` method if possible
If GAP knows the global list `Download_Methods` then
add a new `Download` method in the first place,
such that Julia's `Downloads.download` is used.
This way, we do no longer rely on whether or not some
GAP package or external program is available that provides
download functionality.

This pull request should not be merged at the moment
because it relies on gap-packages/utils/pull/48,
which may change before it gets merged.