admit prescribing a package version ...
... in `GAP.Packages.install`.
This is meanwhile supported by GAP's PackageManager package.

The most interesting application is perhaps `GAP.Packages.load`
with argument `install = true`, because now we can ask for loading a
good enough version of a perhaps previously not installed package,
and get `false` if this is not possible.

(Also fixed now: `GAP.Packages.load` had returned `fail` in some situations,
now `false` is returned in these cases.)