Ensure Ncurses_jll is new enough
Since 6.4.1 it sets TERMINFO_DIRS automatically, so we can get rid of
our workaround for setting it before loading GAP.

That workaround was always incomplete anyway, because while it worked
for GAP's builtin GNU readline support, it is also needed for the
`Browse` packages. But for that it only works if `Browse` is loaded
during GAP startup, which is not always the case: with a fresh GAP.jl
installation, won't have been compiled yet. The user may now try to
compile/install it via `GAP.Packages.install` resp. `GAP.Packages.load`.

All kinds of things can go wrong at that point (including that it may
end up linking against the "wrong").

Note that we do not even load `Ncurses_jll` directly; this is done
indirectly via `GAP_jll`. So the only reason it now appears in
`Project.toml` is to ensure a "new enough" version of it is used. At
some point in the future this should probably instead be moved to
`GAP_jll`.