force version 1.0 of Documenter.jl
(as in AbstractAlgebra.jl and Nemo.jl)

For that, I have changed the three cross-references to
the JuliaInterface manual from relative
(pointing to `../assets/html/JuliaInterface/chap0_mj.html`)
to absolute
(pointing to `https://oscar-system.github.io/GAP.jl/stable/assets/html/JuliaInterface/chap0_mj.html`).

This is not really nice, but I did not find another solution.
When Documenter.jl (in version 1.0 or higher) checks the document's structure,
it does not accept relative links outside the build directory,
and apparently there is no option to suppress this check.

(One could think of encoding the links in question as `http://..` URLs
in the document, and inserting a new step into the build process,
after the checks of course, which just replaces these URLs by the
intended relative ones.  I think this would be absurd.)