adjust the conversion of ranges
After generalizing the three-argument method of `julia_to_gap`
from `Vector` to `AbstractVector`,
the default method that discards the last two arguments is no longer
applicable to `AbstractRange{<:Integer}`,
hence the special one-argument method does not get called.

For the moment, install a three-argument method for `AbstractRange{<:Integer}`.

(And add tests that check whether we really get a GAP `IsRange` object in the end.)