Provide a stale_age to mkpidlock
Normally the pidlock should be held for a short duration; waiting for
its age to exceed the given limits of 10 seconds (for the fake gaproot,
which normally finishes in milliseconds) resp. 300 seconds (for
installing packages, which only takes a few seconds for most, but a few
bad ones can take minutes) until we consider it stale should be safe in
each use case. In fact Julia waits 5 times longer if the process
creating the pid lock file seems to be still running.

On the other hand, without a stale age, the lock file is *never*
considered stale, even if the process creating it definitely is gone,
and so the user can get stuck, which obviously is very bad. To
get unstuck they need to manually delete the lock file.