provide a scratchspace for AtlasRep data if necessary
If AtlasRep does not know a writable directory for downloaded data,
i.e., if the value of its user preference AtlasRepDataDirectory is
an empty string, provide a scratchspace and adjust the user preference.

For that, we check for the empty string when GAP has been started
(which may trigger that AtlasRep gets loaded)
and in `GAP.Packages.load` calls.

Thus this mechanism will fail if one starts GAP without packages
and then loads GAP packages only via the GAP function `LoadPackage`.