fix a test
The tests

```
@test GAP.Packages.install("fga", interactive = false)
@test ! isempty(GAP.Packages.locate_package("fga"))
```

pass only if the FGA package was already loaded before the tests,
which is the case in usual installations.
If one starts GAP without packages (`-A` option) then the second test fails.
One has to load the package in order to be sure that `locate_package`
returns a nonempty result.