Revamp GAP package overrides (#1094)
Previously we only checked for the package name to decide whether to
override the result of DirectoriesPackagePrograms. However, this is
problematic if e.g. we provide a kernel extension for version 1.0 of a
GAP package, but the user tries to load version 2.0 which may have an
incompatible kernel extension.

Therefore the new code is much stricter: we check whether the specific
version of the package for which `DirectoriesPackagePrograms` was called
is the one we provide via an artifact, and only then do we activate the
override.